requirements.txt
fastapi
uvicorn
sqlalchemy
xxhash
aiofiles
python-multipart
Pillow
pywin32
